aws_lambda/.debug/requirements.txt
ikp3db==1.1.4
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
Flask==1.1.2
Flask-SQLAlchemy==2.4.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
lxml==4.5.0
MarkupSafe==1.1.1
mysql-connector-python-rf==2.2.2
numpy==1.18.2
pandas==1.0.3
psycopg2-binary==2.8.5
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
soupsieve==2.0
SQLAlchemy==1.3.15
urllib3==1.25.8
Werkzeug==1.0.1
